# Conflicts:
#	src/main/java/com/mebay/controller/system/DepController.java
#	src/main/resources/application.properties

+++ src/main/java/com/mebay/service/DepartmentService.java
@@ -17,7 +17,7 @@ import java.util.List;
 import java.util.logging.Logger;
 
 @Service
-@Transactional(rollbackFor={Exception.class})
+@Transactional(rollbackFor = {Exception.class})
 public class DepartmentService {
     private static final Logger logger = Logger.getLogger(DepartmentService.class.getSimpleName());
     private final DepartmentMapper departmentMapper;
@@ -31,20 +31,19 @@ public class DepartmentService {
 
     /**
      * 创建一个单位
+     *
      * @param department 单位信息
      */
     public int creationDep(Department department) {
         department.setEnabled(true);
         User user = UserUtils.getCurrentUser();
-        department.setParentId(UserUtils.getCurrentUser().getDepId());
+        department.setParentId(user.getDepId());
         //logger.info(department.getParentId() + " " + (deps == null));
         if (user.getRole().contains(new Role("ROLE_HIGH_GRADE_ADMIN"))) {
-            if (department.getParentId().equals(1L)) {
-                department.setCreationUid(user.getId());
-                logger.info("nmka>>>>>>>>>>>");
-            }
+            department.setCreationUid(user.getId());
+            logger.info("nmka>>>>>>>>>>>");
         }
-        if (departmentMapper.enable(department.getParentId())) {
+        if (departmentMapper.enable(user.getDepId())) {
             return departmentMapper.addDep(department);
         }
         return -1;
@@ -53,35 +52,44 @@ public class DepartmentService {
     public int updateDep(Department dep, Long did) {
         if (Util.isEmpty(dep))
             return -2;
-        List<DeptTreeId> deptTreeIds = getDeptIdTreeByUser();
-        for (DeptTreeId deptId : deptTreeIds) {
-            DeptTreeId d = deptId.findSubById(did);
-            if (d != null) {
-                String logoPath = getDepById(d.getId()).getLogo();
-                int i = departmentMapper.updateDep(dep, did);
-                if (i == 1 && logoPath != null) {
-                    FileUtil.deleteFile("." + logoPath);
-                }
-                return i;
+        Long deptId = UserUtils.getCurrentUser().getDepId();
+        if (!did.equals(deptId)) {  //如果修改的不是自己
+            List<Department> depts = departmentMapper.findDeptByPid(deptId);
+            if (!Util.hasAny((s, d) -> s.getId().equals(d), depts, did)) {  //也不是自己管理的单位下的
+                return -1;
             }
         }
-        return -1;
+        String logoPath = getDepById(did).getLogo();
+        int i = departmentMapper.updateDep(dep, did);
+        logger.info(logoPath + " >> " + dep.getLogo());
+        if (i == 1 && logoPath != null && dep.getLogo() != null && !logoPath.equals(dep.getLogo())) {
+            FileUtil.deleteFile("." + logoPath);
+            logger.info("删除》》》》》");
+        }
+        return i;
     }
 
     /**
      * 获取单位id树通过当前用户。
      * 单位id树：为了节省性能，该方法只返回单位的id
      */
-    public List<DeptTreeId> getDeptIdTreeByUser() {
+    public List<IdTree> getDeptIdTreeByUser() {
         User user = UserUtils.getCurrentUser();
         if (Util.hasAny(Role::equalsRole, user.getRole(), "ROLE_HIGH_GRADE_ADMIN")) {
             return departmentMapper.getDeptsByCreationId(user.getId());
+        }else if (Util.hasAny(Role::equalsRole, user.getRole(), "ROLE_SUPER_ADMIN")) {
+            //TODO
         }
         return Collections.singletonList(departmentMapper.getDeptIdTreeById(user.getDepId()));
     }
 
+    public List<IdTree> getDeptsByCreationId(Long id) {
+        return departmentMapper.getDeptsByCreationId(id);
+    }
+
     /**
      * 获取单位
+     *
      * @return 单位
      */
     public PageView<Department> getDeptByUid(PageQuery pageQuery) {
@@ -97,20 +105,18 @@ public class DepartmentService {
     }
 
     public int deleteDep(Long id) {
-        List<DeptTreeId> deptTreeIds = getDeptIdTreeByUser();
-        for (DeptTreeId deptId : deptTreeIds) {
-            DeptTreeId d = deptId.findSubById(id);
-            if (d != null) {
-                String logoPath = getDepById(d.getId()).getLogo();
-                if(d.getChildren().isEmpty() && userService.getUserByDepId(d.getId()).isEmpty()) {
-                    int i = departmentMapper.deleteDep(id);
-                    if(i == 1) {
-                        FileUtil.deleteFile("." + logoPath);
-                    }
-                    return i;
-                }else
-                    return -2;
-            }
+        List<Department> depts = departmentMapper.findDeptByPid(UserUtils.getCurrentUser().getDepId());
+        if (Util.hasAny((dept, d) -> dept.getId().equals(d), depts, id)) {
+            IdTree d = departmentMapper.getDeptIdTreeById(id);
+            String logoPath = departmentMapper.getDeptById(id).getLogo();
+            if (d.getChildren().isEmpty() && userService.getUserByDepId(d.getId()).isEmpty()) { //如果单位下有子单位或用户
+                int i = departmentMapper.deleteDep(id);
+                if (i == 1) {
+                    FileUtil.deleteFile("." + logoPath);
+                }
+                return i;
+            } else
+                return -2;
         }
         return -1;
     }
